got test cases running
