+++ tools/src/wyvern/tools/tests/ExampleTests.java
@@ -1,5 +1,12 @@
 package wyvern.tools.tests;
 
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -8,7 +15,9 @@ import org.junit.experimental.categories.Category;
 import wyvern.stdlib.Globals;
 import wyvern.target.corewyvernIL.expression.IntegerLiteral;
 import wyvern.target.corewyvernIL.support.Util;
+import wyvern.tools.Interpreter;
 import wyvern.tools.PythonCompiler;
+import wyvern.tools.errors.ToolError;
 import wyvern.tools.imports.extensions.WyvernResolver;
 import wyvern.tools.parsing.coreparser.ParseException;
 import wyvern.tools.tests.suites.RegressionTests;
@@ -51,6 +60,118 @@ public class ExampleTests {
     public void testLinkedList() throws ParseException {
         TestUtil.doTestScriptModularly(PATH, "rosetta.linkedList", Util.intType(),
                 Util.intValue(3));
+    }
+
+    @Test
+    public void testCellModuleClient() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "modules.cellModuleClient",
+          Util.intType(), new IntegerLiteral(2));
+    }
+
+    @Test
+    public void testHello() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "rosetta.hello", Util.unitType(),
+          Util.unitValue());
+    }
+
+    @Test
+    public void testCellClientMain() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "modules.cellClientMain",
+          Util.intType(), new IntegerLiteral(1));
+    }
+
+    @Test
+    public void testCrossPlatformHello() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "xplatform.hello-via-writer",
+          Util.unitType(), Util.unitValue());
+    }
+
+    @Test
+    public void testExplicitCrossPlatformHello() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "xplatform.hello-explicit-writer",
+          Util.unitType(), Util.unitValue());
+    }
+
+    @Test
+    public void testMembrane() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "capabilities.Membrane", null, null);
+    }
+    
+    @Test
+    public void testHelloExplicit() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "rosetta.hello-explicit",
+          Util.unitType(), Util.unitValue());
+    }
+
+    @Test
+    public void testFunctions() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "introductory.functions",
+          Util.intType(), new IntegerLiteral(6));
+    }
+
+    @Test
+    public void testIOLibServerClient() throws ParseException {
+      ExecutorService executor = Executors.newFixedThreadPool(2);
+
+      Future<?> futureServer = executor.submit(() -> {
+        try {
+          TestUtil.doTestScriptModularly(PATH, "io-lib-redesign.server", Util.unitType(),
+              Util.unitValue());
+        } catch (Exception e) {
+          throw new RuntimeException(e);
+        }
+      });
+
+      // We need to let the server start and get to waiting/blocking on a socket
+      // with accept before we start client.
+      // Thus I wait 20 seconds. The following code will also catch any ToolError
+      // that might have happened in the test.
+
+      try {
+        futureServer.get(20, TimeUnit.SECONDS);
+      } catch (InterruptedException e) {
+        // This one is OK.
+      } catch (TimeoutException e) {
+        // This one is a good sign as it is waiting on a socket.
+      } catch (ExecutionException e) {
+        if (e.getCause().getCause() instanceof ToolError) {
+          throw (ToolError) e.getCause().getCause();
+        } else {
+          throw new RuntimeException(e.getCause());
+        }
+      } finally {
+        executor.shutdownNow();
+      }
+
+      // Client can now run and it will complete and shutdown the server as a
+      // result if all goes well.
+      try {
+        TestUtil.doTestScriptModularly(PATH, "io-lib-redesign.client", Util.unitType(),
+            Util.unitValue());
+      } finally {
+        // Just in case - we need to clean up.
+        executor.shutdown();
+        try {
+          if (!executor.awaitTermination(3, TimeUnit.SECONDS)) {
+            executor.shutdownNow();
+          }
+        } catch (InterruptedException e) {
+        } finally {
+          executor.shutdownNow();
+        }
+      }
+    }
+
+    @Test
+    public void testCellClient() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "modules.cellClient", Util.intType(),
+          new IntegerLiteral(7));
+    }
+
+    @Test
+    public void testJavaFFI() throws ParseException {
+      TestUtil.doTestScriptModularly(PATH, "ffi.callFromJava", Util.unitType(),
+          Util.unitValue());
     }
 
     @Test
@@ -166,7 +287,8 @@ public class ExampleTests {
 
     @Test
     public void testCommandLineArguments() throws ParseException {
-        TestUtil.doTestScriptModularly(PATH, "commandLineArguments", Util.unitType(),
-        Util.unitValue());
+        String[] args = new String[] {TestUtil.EXAMPLES_PATH + "commandLineArguments.wyv"};
+        Interpreter.wyvernHome.set("..");
+        Interpreter.main(args);
     }
 }
