+++ tools/src/wyvern/tools/tests/ExamplesNoPrelude.java
@@ -1,12 +1,5 @@
 package wyvern.tools.tests;
 
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-
 import org.hamcrest.core.StringContains;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -57,18 +50,6 @@ public class ExamplesNoPrelude {
 
 
   @Test
-  public void testHello() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "rosetta.hello", Util.unitType(),
-        Util.unitValue());
-  }
-
-  @Test
-  public void testHelloExplicit() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "rosetta.hello-explicit",
-        Util.unitType(), Util.unitValue());
-  }
-
-  @Test
   public void testDatatypes() throws ParseException {
     TestUtil.doTestScriptModularly(PATH, "introductory.datatypes",
         Util.stringType(), new StringLiteral("(x => x) unit"));
@@ -96,12 +77,6 @@ public class ExamplesNoPrelude {
   }
 
   @Test
-  public void testFunctions() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "introductory.functions",
-        Util.intType(), new IntegerLiteral(6));
-  }
-
-  @Test
   public void testObjects() throws ParseException {
     TestUtil.doTestScriptModularly(PATH, "introductory.objects", Util.intType(),
         new IntegerLiteral(7));
@@ -126,33 +101,9 @@ public class ExamplesNoPrelude {
   }
 
   @Test
-  public void testCellClient() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "modules.cellClient", Util.intType(),
-        new IntegerLiteral(7));
-  }
-
-  @Test
   public void testMaybe() throws ParseException {
     TestUtil.doTestScriptModularly(PATH, "introductory.maybe", Util.intType(),
         new IntegerLiteral(15));
-  }
-
-  @Test
-  public void testCellModuleClient() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "modules.cellModuleClient",
-        Util.intType(), new IntegerLiteral(2));
-  }
-
-  @Test
-  public void testCellClientMain() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "modules.cellClientMain",
-        Util.intType(), new IntegerLiteral(1));
-  }
-
-  @Test
-  public void testJavaFFI() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "ffi.callFromJava", Util.unitType(),
-        Util.unitValue());
   }
 
   @Test
@@ -161,23 +112,6 @@ public class ExamplesNoPrelude {
         Util.unitType(), Util.unitValue());
   }
 
-  @Test
-  public void testCrossPlatformHello() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "xplatform.hello-via-writer",
-        Util.unitType(), Util.unitValue());
-  }
-
-  @Test
-  public void testExplicitCrossPlatformHello() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "xplatform.hello-explicit-writer",
-        Util.unitType(), Util.unitValue());
-  }
-
-  @Test
-  public void testMembrane() throws ParseException {
-    TestUtil.doTestScriptModularly(PATH, "capabilities.Membrane", null, null);
-  }
-  
   @Test
   public void testCodeCompletion() throws ParseException {
     TestUtil.doTestScriptModularly(PATH, "effects.codeCompletion", null, null);
@@ -216,59 +150,6 @@ public class ExamplesNoPrelude {
   @Test
   public void testTwice() throws ParseException {
     TestUtil.doTestScriptModularly(PATH, "effects.twice", null, null);
-  }
-
-  @Test
-  public void testIOLibServerClient() throws ParseException {
-    ExecutorService executor = Executors.newFixedThreadPool(2);
-
-    Future<?> futureServer = executor.submit(() -> {
-      try {
-        TestUtil.doTestScriptModularly(PATH, "io-lib.server", Util.unitType(),
-            Util.unitValue());
-      } catch (Exception e) {
-        throw new RuntimeException(e);
-      }
-    });
-
-    // We need to let the server start and get to waiting/blocking on a socket
-    // with accept before we start client.
-    // Thus I wait 3 seconds. The following code will also catch any ToolError
-    // that might have happened in the test.
-
-    try {
-      futureServer.get(3, TimeUnit.SECONDS);
-    } catch (InterruptedException e) {
-      // This one is OK.
-    } catch (TimeoutException e) {
-      // This one is a good sign as it is waiting on a socket.
-    } catch (ExecutionException e) {
-      if (e.getCause().getCause() instanceof ToolError) {
-        throw (ToolError) e.getCause().getCause();
-      } else {
-        throw new RuntimeException(e.getCause());
-      }
-    } finally {
-      executor.shutdownNow();
-    }
-
-    // Client can now run and it will complete and shutdown the server as a
-    // result if all goes well.
-    try {
-      TestUtil.doTestScriptModularly(PATH, "io-lib.client", Util.unitType(),
-          Util.unitValue());
-    } finally {
-      // Just in case - we need to clean up.
-      executor.shutdown();
-      try {
-        if (!executor.awaitTermination(3, TimeUnit.SECONDS)) {
-          executor.shutdownNow();
-        }
-      } catch (InterruptedException e) {
-      } finally {
-        executor.shutdownNow();
-      }
-    }
   }
 
 }
